ontokit test separate
