+++ requirements_ontokit.txt
@@ -1,11 +1,7 @@
-myst-parser>=4.0,<4.1; python_version<'3.11'
-myst-parser>=5.0,<5.1; python_version>='3.11'
-nbsphinx>=0.9.0,<0.10
-pydata-sphinx-theme>=0.16.1,<0.17
-sphinx>=7.4,<8.2; python_version<'3.11'
-sphinx>=8.2,<9.0; python_version>='3.11' and python_version<'3.12'
-sphinx>=9.1,<9.2; python_version>='3.12'
-sphinx_design>=0.6.1,<0.7; python_version<'3.11'
-sphinx_design>=0.7.0,<0.8; python_version>='3.11'
-sphinxcontrib-globalsubs==0.1.2; python_version<'3.12'
-sphinxcontrib-globalsubs>=1.1.0,<1.2; python_version>='3.12'
+myst-parser>=5.0.0;python_version>='3.12'
+myst-parser<5.0.0;python_version<'3.12'
+nbsphinx==0.9.8
+pydata-sphinx-theme==0.16.1
+sphinx==9.1.0
+sphinx_design==0.7.0
+sphinxcontrib-globalsubs==1.0.0
